# Optimize IsIntegrationEnabled