--- a/src/Datadog.Trace.ClrProfiler.Managed/Integrations/AdoNet/SqlTags.cs
+++ b/src/Datadog.Trace.ClrProfiler.Managed/Integrations/AdoNet/SqlTags.cs
@@ -18,7 +18,7 @@ namespace Datadog.Trace.ClrProfiler.Integrations.AdoNet
 
         public string DbType { get; set; }
 
-        public string InstrumentationName => IntegrationRegistry.Names[AdoNetConstants.IntegrationId];
+        public string InstrumentationName => IntegrationRegistry.GetName(AdoNetConstants.IntegrationId);
 
         public string DbName { get; set; }