--- a/src/Datadog.Trace.ClrProfiler.Managed/Integrations/GraphQLTags.cs
+++ b/src/Datadog.Trace.ClrProfiler.Managed/Integrations/GraphQLTags.cs
@@ -16,7 +16,7 @@ namespace Datadog.Trace.ClrProfiler.Integrations
 
         public override string SpanKind => SpanKinds.Server;
 
-        public string InstrumentationName => IntegrationRegistry.Names[GraphQLIntegration.IntegrationId];
+        public string InstrumentationName => IntegrationRegistry.GetName(GraphQLIntegration.IntegrationId);
 
         public string Language => TracerConstants.Language;