… src/Datadog.Trace.ClrProfiler.Managed/Integrations/MongoDbTags.cs
… src/Datadog.Trace.ClrProfiler.Managed/Integrations/RedisHelper.cs
… src/Datadog.Trace.ClrProfiler.Managed/Integrations/ServiceStackRedisIntegration.cs
… src/Datadog.Trace.ClrProfiler.Managed/Integrations/StackExchange.Redis/ConnectionMultiplexer.cs
… src/Datadog.Trace.ClrProfiler.Managed/Integrations/StackExchange.Redis/RedisBatch.cs
… src/Datadog.Trace.ClrProfiler.Managed/Integrations/Testing/XUnitIntegration.cs
… src/Datadog.Trace.ClrProfiler.Managed/Integrations/WcfIntegration.cs
… src/Datadog.Trace.ClrProfiler.Managed/Integrations/WebRequestIntegration.cs
… src/Datadog.Trace.ClrProfiler.Managed/ScopeFactory.cs
… src/Datadog.Trace.ServiceFabric/Remoting.cs
… src/Datadog.Trace/Configuration/IntegrationIds.cs
… src/Datadog.Trace/Configuration/IntegrationInfo.cs
… src/Datadog.Trace/Configuration/IntegrationRegistry.cs
… src/Datadog.Trace/Configuration/IntegrationSettingsCollection.cs
… src/Datadog.Trace/Configuration/TracerSettings.cs
… src/Datadog.Trace/DiagnosticListeners/AspNetCoreDiagnosticObserver.cs
… src/Datadog.Trace/Tagging/InstrumentationTags.cs
… src/Datadog.Trace/Tracer.cs
… test/Datadog.Trace.ClrProfiler.Managed.Tests/ScopeFactoryTests.cs