Index integrations by id instead of name for faster lookup time.

|                   Method |           Job |       Runtime |      Mean |     Error |    StdDev | Ratio | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------- |-------------- |-------------- |----------:|----------:|----------:|------:|------:|------:|------:|----------:|
| IsIntegrationEnabled_Old |    .NET 4.7.2 |    .NET 4.7.2 | 64.977 ns | 0.6788 ns | 0.6017 ns |  1.00 |     - |     - |     - |         - |
| IsIntegrationEnabled_New |    .NET 4.7.2 |    .NET 4.7.2 |  6.803 ns | 0.0439 ns | 0.0411 ns |  0.10 |     - |     - |     - |         - |
|                          |               |               |           |           |           |       |       |       |       |           |
| IsIntegrationEnabled_Old | .NET Core 3.1 | .NET Core 3.1 | 42.046 ns | 0.4016 ns | 0.3354 ns |  1.00 |     - |     - |     - |         - |
| IsIntegrationEnabled_New | .NET Core 3.1 | .NET Core 3.1 |  6.620 ns | 0.0402 ns | 0.0356 ns |  0.16 |     - |     - |     - |         - |

I expect the gain to be bigger on multithreaded scenarios.

Note: **`TracerSettings.Integrations` is public, so this is technically a breaking change.** I doubt people used it though.