@kevingosse 2020-11-05
I just realized something else while fixing ServiceRemoting: with that change, it's not possible to write an integration without updating Datadog.Trace. If that's a problem, maybe I can use a hybrid approach:
 - Use id-based indexing for code that is aware of it
 - Expose the legacy name-based system for others

It would increase the complexity a bit, so tell me if you think it's worth the trouble.