@kevingosse 2020-11-05
I ended up making the change.

The tricky bit is with the DisabledIntegrations collection. Unfortunately it can be modified past the creation of the TracerSettings object. As I didn't want to keep querying this collection every time `IsIntegrationEnabled` is called, I introduced a `TracerSettings.Freeze` that is call by the tracer during initialization. When it's called, the list of disabled integrations is merged with the integration settings, and modifications past this point won't be taken into account.

Modifications to DisabledIntegrations after creating the tracer weren't supported anyway since there's no synchronization when accessing the collection.