@lucaspimentel 2020-11-05
> Note: TracerSettings.Integrations is public, so this is technically a breaking change. I doubt people used it though.

I think maybe this isn't true anymore since you added back the legacy string-base API.